feat: migrate xueyi
